refactor(api): split admin route monolith into sub-package
Split observal-server/api/routes/admin.py (1518 lines) into focused modules:

- admin/helpers.py: branding validators, _generate_unique_password, _get_user_org
- admin/enterprise_settings.py: diagnostics, system warnings, settings CRUD, resources/apply
- admin/users.py: user management (list, create, role, department, password, delete)
- admin/penalties.py: penalty definitions, dimension weights, canary config
- admin/org.py: security events, trace privacy, registered agents only, cache
- admin/retention.py: data retention config, preview, stats, warnings

Uses shared _router.py pattern so main.py requires no import changes.
Test mock.patch targets updated to reference new sub-module paths.
Migrated from settings.* to services.dynamic_settings (ds) to match upstream.
