--- observal-server/api/routes/admin.py
@@ -1,1566 +1,0 @@
-# SPDX-FileCopyrightText: 2026 Subramania Raja <dhanpraja231@gmail.com>
-# SPDX-FileCopyrightText: 2026 Hari Srinivasan <harisrini21@gmail.com>
-# SPDX-FileCopyrightText: 2026 Kaushik Kumar <kaushikrjpm10@gmail.com>
-# SPDX-FileCopyrightText: 2026 Lokesh Selvam <lokeshselvam7025@gmail.com>
-# SPDX-FileCopyrightText: 2026 Shaan Narendran <shaannaren06@gmail.com>
-# SPDX-FileCopyrightText: 2026 Shreem Seth <shreemseth26@gmail.com>
-# SPDX-FileCopyrightText: 2026 Swathi Saravanan <ss4522@cornell.edu>
-# SPDX-FileCopyrightText: 2026 Vishnu Muthiah <vishnu.muthiah04@gmail.com>
-# SPDX-License-Identifier: AGPL-3.0-only
-
-import base64
-import hashlib
-import json
-import logging
-import re
-import secrets
-import uuid
-
-from fastapi import APIRouter, Depends, HTTPException
-from pydantic import BaseModel
-from redis.exceptions import RedisError
-from sqlalchemy import func, select, text
-from sqlalchemy.exc import IntegrityError
-from sqlalchemy.ext.asyncio import AsyncSession
-
-import services.dynamic_settings as ds
-from api.deps import ROLE_HIERARCHY, get_db, get_or_create_default_org, require_password_auth, require_role
-from config import settings
-from models.enterprise_config import EnterpriseConfig
-from models.organization import Organization
-from models.user import User, UserRole
-from schemas.admin import (
-    AdminResetPasswordRequest,
-    EnterpriseConfigResponse,
-    EnterpriseConfigUpdate,
-    UserAdminResponse,
-    UserCreateRequest,
-    UserCreateResponse,
-    UserDepartmentUpdate,
-    UserRoleUpdate,
-)
-from schemas.retention import RetentionConfigResponse, RetentionConfigUpdate
-from services.audit_helpers import audit
-from services.security_events import (
-    EventType,
-    SecurityEvent,
-    Severity,
-    emit_security_event,
-)
-from services.username_generator import generate_unique_username
-
-logger = logging.getLogger(__name__)
-
-router = APIRouter(prefix="/api/v1/admin", tags=["admin"])
-
-
-# ── Diagnostics ─────────────────────────────────────────
-
-
-@router.get("/diagnostics")
-async def diagnostics(
-    db: AsyncSession = Depends(get_db),
-    current_user: User = Depends(require_role(UserRole.admin)),
-):
-    """Authenticated system health: full status for ops dashboards."""
-    from services.crypto import get_key_manager
-
-    deployment_mode = settings.DEPLOYMENT_MODE
-    diag: dict[str, object] = {
-        "status": "ok",
-        "deployment_mode": deployment_mode,
-        "checks": {},
-    }
-
-    # Database
-    try:
-        await db.execute(text("SELECT 1"))
-        user_count = await db.scalar(select(func.count()).select_from(User))
-        demo_count = await db.scalar(select(func.count()).select_from(User).where(User.is_demo.is_(True)))
-        diag["checks"]["database"] = {
-            "status": "ok",
-            "users": user_count or 0,
-            "demo_accounts": demo_count or 0,
-        }
-    except Exception as e:
-        diag["checks"]["database"] = {"status": "error", "detail": str(e)}
-        diag["status"] = "unhealthy"
-
-    # JWT keys
-    try:
-        get_key_manager()
-        diag["checks"]["jwt_keys"] = {
-            "status": "ok",
-            "algorithm": settings.JWT_SIGNING_ALGORITHM,
-        }
-    except RuntimeError:
-        diag["checks"]["jwt_keys"] = {
-            "status": "missing",
-            "algorithm": settings.JWT_SIGNING_ALGORITHM,
-        }
-
-    # Enterprise config
-    if deployment_mode == "enterprise":
-        issues: list[str] = []
-        if settings.SECRET_KEY == "change-me-to-a-random-string":
-            issues.append("SECRET_KEY is using default value")
-        sso_only = await ds.get_bool("deployment.sso_only")
-        frontend_url = await ds.get("deployment.frontend_url")
-        if sso_only and not settings.OAUTH_CLIENT_ID:
-            issues.append("OAUTH_CLIENT_ID is not set (required for SSO-only mode)")
-        if frontend_url in ("http://localhost:3000", ""):
-            issues.append("deployment.frontend_url is localhost")
-        diag["checks"]["enterprise"] = {
-            "status": "ok" if not issues else "misconfigured",
-            "sso_only": sso_only,
-            "sso_configured": bool(settings.OAUTH_CLIENT_ID),
-            "issues": issues,
-        }
-        if issues:
-            diag["status"] = "degraded"
-
-    await audit(current_user, "admin.diagnostics.view", "diagnostics")
-    return diag
-
-
-# ── Branding Validation ──────────────────────────────────
-
-_ALLOWED_LOGO_MIMES = {
-    "image/png",
-    "image/svg+xml",
-    "image/x-icon",
-    "image/vnd.microsoft.icon",
-    "image/jpeg",
-    "image/webp",
-}
-_MAX_LOGO_BYTES = 2 * 1024 * 1024
-_MAX_DATA_URL_LEN = 3 * 1024 * 1024  # base64 bloats ~33%, cap raw string
-_MAX_APP_NAME_LEN = 30
-
-# Magic byte signatures for allowed image formats
-_MAGIC_BYTES: dict[str, list[bytes]] = {
-    "image/png": [b"\x89PNG\r\n\x1a\n"],
-    "image/jpeg": [b"\xff\xd8\xff"],
-    "image/webp": [b"RIFF"],  # RIFF....WEBP
-    "image/x-icon": [b"\x00\x00\x01\x00", b"\x00\x00\x02\x00"],
-    "image/vnd.microsoft.icon": [b"\x00\x00\x01\x00", b"\x00\x00\x02\x00"],
-    # SVG is validated separately via _sanitize_svg
-}
-
-# SVG elements and attributes that can execute code or make external requests
-_SVG_DANGEROUS_TAGS = re.compile(
-    r"<[\s/]*(script|foreignObject|iframe|embed|object|applet|meta|link|style|handler|set|animate|animateTransform|animateMotion)\b",
-    re.IGNORECASE,
-)
-_SVG_EVENT_ATTRS = re.compile(r"\bon\w+\s*=", re.IGNORECASE)
-_SVG_JS_HREF = re.compile(r"(?:href|xlink:href)[\s=\"']*javascript:", re.IGNORECASE)
-_SVG_EXTERNAL_REF = re.compile(r"(?:href|xlink:href|src|url)[\s=\"']*(?:https?://|//|data:(?!image/))", re.IGNORECASE)
-_SVG_XML_DECL = re.compile(r"<!(?:DOCTYPE|ENTITY)\b", re.IGNORECASE)
-
-# Characters forbidden in the app name
-_UNSAFE_NAME_CHARS = re.compile("[\x00-\x1f\x7f\u200b-\u200f\u202a-\u202e\u2060-\u2064\ufeff]")
-
-
-def _validate_magic_bytes(raw: bytes, mime_type: str) -> None:
-    """Verify the file's actual bytes match the claimed MIME type."""
-    signatures = _MAGIC_BYTES.get(mime_type)
-    if signatures is None:
-        return  # SVG handled separately
-    if not any(raw.startswith(sig) for sig in signatures):
-        raise HTTPException(status_code=422, detail=f"File content does not match declared type {mime_type}")
-    if mime_type == "image/webp" and raw[8:12] != b"WEBP":
-        raise HTTPException(status_code=422, detail="File content does not match declared type image/webp")
-
-
-def _sanitize_svg(raw: bytes) -> bytes:
-    """Reject SVGs containing dangerous elements or attributes."""
-    try:
-        text = raw.decode("utf-8")
-    except UnicodeDecodeError:
-        raise HTTPException(status_code=422, detail="SVG contains invalid UTF-8")
-
-    if _SVG_XML_DECL.search(text):
-        raise HTTPException(status_code=422, detail="SVG must not contain DOCTYPE or ENTITY declarations")
-    if _SVG_DANGEROUS_TAGS.search(text):
-        raise HTTPException(
-            status_code=422, detail="SVG contains forbidden elements (script, foreignObject, iframe, etc.)"
-        )
-    if _SVG_EVENT_ATTRS.search(text):
-        raise HTTPException(
-            status_code=422, detail="SVG contains forbidden event handler attributes (onclick, onload, etc.)"
-        )
-    if _SVG_JS_HREF.search(text):
-        raise HTTPException(status_code=422, detail="SVG contains javascript: URLs")
-    if _SVG_EXTERNAL_REF.search(text):
-        raise HTTPException(status_code=422, detail="SVG contains external resource references")
-
-    return raw
-
-
-def _validate_branding_logo(value: str) -> None:
-    if not value:
-        return
-
-    if len(value) > _MAX_DATA_URL_LEN:
-        raise HTTPException(status_code=422, detail="Image data too large")
-
-    match = re.match(r"^data:(image/[a-zA-Z0-9.+-]+);base64,(.+)$", value, re.DOTALL)
-    if not match:
-        raise HTTPException(status_code=422, detail="Logo must be a base64 data URL (data:image/...;base64,...)")
-    mime_type = match.group(1)
-    b64_data = match.group(2)
-    if mime_type not in _ALLOWED_LOGO_MIMES:
-        raise HTTPException(
-            status_code=422, detail=f"Unsupported image type: {mime_type}. Allowed: PNG, SVG, ICO, JPEG, WEBP"
-        )
-    try:
-        raw = base64.b64decode(b64_data)
-    except Exception:
-        raise HTTPException(status_code=422, detail="Invalid base64 data")
-    if len(raw) > _MAX_LOGO_BYTES:
-        size_mb = round(len(raw) / (1024 * 1024), 1)
-        max_mb = _MAX_LOGO_BYTES // (1024 * 1024)
-        raise HTTPException(status_code=422, detail=f"Logo too large ({size_mb}MB). Maximum: {max_mb}MB")
-
-    if mime_type == "image/svg+xml":
-        _sanitize_svg(raw)
-    else:
-        _validate_magic_bytes(raw, mime_type)
-
-
-def _validate_branding_app_name(value: str) -> None:
-    if len(value) > _MAX_APP_NAME_LEN:
-        raise HTTPException(
-            status_code=422, detail=f"App name too long ({len(value)} chars). Maximum: {_MAX_APP_NAME_LEN}"
-        )
-    if _UNSAFE_NAME_CHARS.search(value):
-        raise HTTPException(status_code=422, detail="App name contains forbidden control or invisible characters")
-    if "<" in value and ">" in value:
-        raise HTTPException(status_code=422, detail="App name must not contain HTML tags")
-
-
-# ── Enterprise Settings ──────────────────────────────────
-
-
-@router.get("/system-warnings", response_model=list[dict])
-async def system_warnings(
-    db: AsyncSession = Depends(get_db),
-    current_user: User = Depends(require_role(UserRole.admin)),
-):
-    """Return actionable security warnings for the admin settings page."""
-    warnings: list[dict] = []
-
-    weak_keys = {"change-me-to-a-random-string", "changeme", "secret", "dev", ""}
-    if settings.SECRET_KEY in weak_keys or len(settings.SECRET_KEY) < 32:
-        warnings.append(
-            {
-                "level": "critical",
-                "code": "weak_secret_key",
-                "message": "SECRET_KEY is insecure. Set a random string of at least 32 characters.",
-            }
-        )
-
-    demo_count = await db.scalar(select(func.count()).select_from(User).where(User.is_demo.is_(True)))
-    if demo_count:
-        warnings.append(
-            {
-                "level": "warning",
-                "code": "demo_accounts_active",
-                "message": f"{demo_count} demo account(s) are still active. Remove them or change their passwords before going to production.",
-            }
-        )
-
-    return warnings
-
-
-@router.get("/settings", response_model=list[EnterpriseConfigResponse])
-async def list_settings(
-    db: AsyncSession = Depends(get_db),
-    current_user: User = Depends(require_role(UserRole.admin)),
-):
-    from services.dynamic_settings import SENSITIVE_KEYS, decrypt_value, mask_value
-
-    result = await db.execute(select(EnterpriseConfig).order_by(EnterpriseConfig.key))
-    configs = []
-    for c in result.scalars().all():
-        resp = EnterpriseConfigResponse.model_validate(c)
-        # Decrypt then mask sensitive values for display
-        if resp.key in SENSITIVE_KEYS:
-            decrypted = decrypt_value(resp.value)
-            resp.value = mask_value(resp.key, decrypted)
-        configs.append(resp)
-    await audit(current_user, "admin.settings.list", "settings")
-    return configs
-
-
-@router.get("/settings/schema")
-async def get_settings_schema(
-    current_user: User = Depends(require_role(UserRole.admin)),
-):
-    """Return the settings schema: section definitions, keys, descriptions, and metadata.
-
-    Used by the frontend to render the settings page with proper grouping,
-    danger-zone styling, and sensitive field masking.
-    """
-    from services.dynamic_settings import DEFAULTS, SECTIONS, SENSITIVE_KEYS
-
-    return {
-        "sections": SECTIONS,
-        "sensitive_keys": list(SENSITIVE_KEYS),
-        "defaults": DEFAULTS,
-    }
-
-
-@router.get("/settings/{key}", response_model=EnterpriseConfigResponse)
-async def get_setting(
-    key: str,
-    db: AsyncSession = Depends(get_db),
-    current_user: User = Depends(require_role(UserRole.admin)),
-):
-    result = await db.execute(select(EnterpriseConfig).where(EnterpriseConfig.key == key))
-    cfg = result.scalar_one_or_none()
-    if not cfg:
-        raise HTTPException(status_code=404, detail="Setting not found")
-    await audit(current_user, "admin.settings.view", "settings", resource_name=key)
-    return EnterpriseConfigResponse.model_validate(cfg)
-
-
-@router.put("/settings/{key}", response_model=EnterpriseConfigResponse)
-async def upsert_setting(
-    key: str,
-    req: EnterpriseConfigUpdate,
-    db: AsyncSession = Depends(get_db),
-    current_user: User = Depends(require_role(UserRole.super_admin)),
-):
-    if key in ("branding.logo", "branding.wordmark"):
-        _validate_branding_logo(req.value)
-    elif key == "branding.app_name":
-        _validate_branding_app_name(req.value)
-
-    # Encrypt sensitive values before storing
-    from services.dynamic_settings import SENSITIVE_KEYS, encrypt_value
-
-    store_value = encrypt_value(req.value) if key in SENSITIVE_KEYS and req.value else req.value
-
-    result = await db.execute(select(EnterpriseConfig).where(EnterpriseConfig.key == key))
-    cfg = result.scalar_one_or_none()
-    if cfg:
-        cfg.value = store_value
-    else:
-        cfg = EnterpriseConfig(key=key, value=store_value)
-        db.add(cfg)
-    await db.commit()
-    await db.refresh(cfg)
-
-    # Invalidate settings cache
-
-    await ds.invalidate(key)
-    await ds.refresh_sync_cache()
-
-    await emit_security_event(
-        SecurityEvent(
-            event_type=EventType.SETTING_CHANGED,
-            severity=Severity.WARNING,
-            outcome="success",
-            actor_id=str(current_user.id),
-            actor_email=current_user.email,
-            actor_role=current_user.role.value,
-            target_id=key,
-            target_type="setting",
-        )
-    )
-    await audit(current_user, "admin.settings.update", "settings", resource_name=key)
-    return EnterpriseConfigResponse.model_validate(cfg)
-
-
-@router.delete("/settings/{key}")
-async def delete_setting(
-    key: str,
-    db: AsyncSession = Depends(get_db),
-    current_user: User = Depends(require_role(UserRole.super_admin)),
-):
-    result = await db.execute(select(EnterpriseConfig).where(EnterpriseConfig.key == key))
-    cfg = result.scalar_one_or_none()
-    if not cfg:
-        raise HTTPException(status_code=404, detail="Setting not found")
-    await db.delete(cfg)
-    await db.commit()
-
-    # Invalidate settings cache
-
-    await ds.invalidate(key)
-    await ds.refresh_sync_cache()
-
-    await audit(current_user, "admin.settings.delete", "settings", resource_name=key)
-    return {"deleted": key}
-
-
-# ── User Management ──────────────────────────────────────
-
-
-@router.get("/users", response_model=list[UserAdminResponse])
-async def list_users(
-    db: AsyncSession = Depends(get_db),
-    current_user: User = Depends(require_role(UserRole.admin)),
-):
-    stmt = select(User).order_by(User.created_at.desc())
-    if current_user.org_id is not None:
-        stmt = stmt.where(User.org_id == current_user.org_id)
-    result = await db.execute(stmt)
-    users = [UserAdminResponse.model_validate(u) for u in result.scalars().all()]
-    await audit(current_user, "admin.users.list", "user")
-    return users
-
-
-@router.post("/users", response_model=UserCreateResponse, dependencies=[Depends(require_password_auth)])
-async def create_user(
-    req: UserCreateRequest,
-    db: AsyncSession = Depends(get_db),
-    current_user: User = Depends(require_role(UserRole.admin)),
-):
-    """Admin creates a new user and gets back their generated password."""
-    existing = await db.execute(select(User).where(User.email == req.email))
-    if existing.scalar_one_or_none():
-        raise HTTPException(status_code=409, detail="Email already registered")
-
-    try:
-        role = UserRole(req.role)
-    except ValueError:
-        raise HTTPException(status_code=422, detail=f"Invalid role. Must be one of: {[r.value for r in UserRole]}")
-
-    if ROLE_HIERARCHY.get(role, 999) < ROLE_HIERARCHY[current_user.role]:
-        raise HTTPException(status_code=403, detail="Cannot assign a role higher than your own")
-
-    password = req.password or await _generate_unique_password(db)
-
-    org_id = current_user.org_id
-    if not org_id:
-        default_org = await get_or_create_default_org(db)
-        org_id = default_org.id
-
-    username = req.username or await generate_unique_username(req.email, db)
-    user = User(email=req.email, username=username, name=req.name, role=role, org_id=org_id)
-    user.set_password(password)
-    db.add(user)
-    try:
-        await db.commit()
-    except IntegrityError:
-        await db.rollback()
-        raise HTTPException(status_code=409, detail="Email already registered")
-    await db.refresh(user)
-
-    await emit_security_event(
-        SecurityEvent(
-            event_type=EventType.USER_CREATED,
-            severity=Severity.INFO,
-            outcome="success",
-            actor_id=str(current_user.id),
-            actor_email=current_user.email,
-            actor_role=current_user.role.value,
-            target_id=str(user.id),
-            target_type="user",
-            detail=f"Created user {user.email} with role {role.value}",
-        )
-    )
-    await audit(
-        current_user,
-        "admin.users.create",
-        "user",
-        resource_id=str(user.id),
-        resource_name=user.email,
-        detail=json.dumps({"role": role.value}),
-    )
-    return UserCreateResponse(
-        id=user.id,
-        email=user.email,
-        username=user.username,
-        name=user.name,
-        role=user.role.value,
-        password=password,
-    )
-
-
-@router.put("/users/{user_id}/role", response_model=UserAdminResponse)
-async def update_user_role(
-    user_id: uuid.UUID,
-    req: UserRoleUpdate,
-    db: AsyncSession = Depends(get_db),
-    current_user: User = Depends(require_role(UserRole.admin)),
-):
-    try:
-        new_role = UserRole(req.role)
-    except ValueError:
-        raise HTTPException(status_code=422, detail=f"Invalid role. Must be one of: {[r.value for r in UserRole]}")
-
-    from api.deps import ROLE_HIERARCHY
-
-    if ROLE_HIERARCHY.get(new_role, 999) < ROLE_HIERARCHY[current_user.role]:
-        raise HTTPException(status_code=403, detail="Cannot assign a role higher than your own")
-
-    if user_id == current_user.id and new_role != current_user.role:
-        raise HTTPException(status_code=400, detail="Cannot change your own role")
-
-    stmt = select(User).where(User.id == user_id)
-    if current_user.org_id is not None:
-        stmt = stmt.where(User.org_id == current_user.org_id)
-    result = await db.execute(stmt)
-    user = result.scalar_one_or_none()
-    if not user:
-        raise HTTPException(status_code=404, detail="User not found")
-    old_role = user.role.value
-    user.role = new_role
-    await db.commit()
-    await db.refresh(user)
-    await emit_security_event(
-        SecurityEvent(
-            event_type=EventType.ROLE_CHANGED,
-            severity=Severity.WARNING,
-            outcome="success",
-            actor_id=str(current_user.id),
-            actor_email=current_user.email,
-            actor_role=current_user.role.value,
-            target_id=str(user.id),
-            target_type="user",
-            detail=f"Role changed from {old_role} to {new_role.value}",
-        )
-    )
-    await audit(
-        current_user,
-        "admin.users.role_update",
-        "user",
-        resource_id=str(user.id),
-        resource_name=user.email,
-        detail=json.dumps({"old_role": old_role, "new_role": new_role.value}),
-    )
-    return UserAdminResponse.model_validate(user)
-
-
-@router.put("/users/{user_id}/department", response_model=UserAdminResponse)
-async def update_user_department(
-    user_id: uuid.UUID,
-    req: UserDepartmentUpdate,
-    db: AsyncSession = Depends(get_db),
-    current_user: User = Depends(require_role(UserRole.admin)),
-):
-    stmt = select(User).where(User.id == user_id)
-    if current_user.org_id is not None:
-        stmt = stmt.where(User.org_id == current_user.org_id)
-    result = await db.execute(stmt)
-    user = result.scalar_one_or_none()
-    if not user:
-        raise HTTPException(status_code=404, detail="User not found")
-    user.department = req.department
-    await db.commit()
-    await db.refresh(user)
-    return UserAdminResponse.model_validate(user)
-
-
-class BulkDepartmentEntry(BaseModel):
-    email: str
-    department: str
-
-
-class BulkDepartmentRequest(BaseModel):
-    entries: list[BulkDepartmentEntry]
-
-
-class BulkDepartmentResult(BaseModel):
-    updated: int
-    not_found: list[str]
-
-
-@router.post("/users/bulk-department", response_model=BulkDepartmentResult)
-async def bulk_update_departments(
-    req: BulkDepartmentRequest,
-    db: AsyncSession = Depends(get_db),
-    current_user: User = Depends(require_role(UserRole.admin)),
-):
-    """Bulk-assign departments to users by email."""
-    updated = 0
-    not_found = []
-
-    for entry in req.entries:
-        stmt = select(User).where(User.email == entry.email.strip().lower())
-        if current_user.org_id is not None:
-            stmt = stmt.where(User.org_id == current_user.org_id)
-        result = await db.execute(stmt)
-        user = result.scalar_one_or_none()
-        if user:
-            user.department = entry.department.strip()
-            updated += 1
-        else:
-            not_found.append(entry.email)
-
-    await db.commit()
-    return BulkDepartmentResult(updated=updated, not_found=not_found)
-
-
-@router.put("/users/{user_id}/password", dependencies=[Depends(require_password_auth)])
-async def reset_user_password(
-    user_id: uuid.UUID,
-    req: AdminResetPasswordRequest,
-    db: AsyncSession = Depends(get_db),
-    current_user: User = Depends(require_role(UserRole.admin)),
-):
-    """Admin resets a user's password.
-
-    Either provide new_password directly, or set generate=true to create
-    a secure random password that doesn't collide with existing hashes.
-    """
-    stmt = select(User).where(User.id == user_id)
-    if current_user.org_id is not None:
-        stmt = stmt.where(User.org_id == current_user.org_id)
-    result = await db.execute(stmt)
-    user = result.scalar_one_or_none()
-    if not user:
-        raise HTTPException(status_code=404, detail="User not found")
-
-    if req.generate:
-        new_password = await _generate_unique_password(db)
-    elif req.new_password:
-        new_password = req.new_password
-    else:
-        raise HTTPException(status_code=422, detail="Provide new_password or set generate=true")
-
-    user.set_password(new_password)
-    await db.commit()
-
-    try:
-        from services.redis import get_redis
-
-        redis = get_redis()
-        await redis.setex(f"must_change_password:{user.id}", 86400, "1")
-    except (RedisError, Exception):
-        pass
-
-    await emit_security_event(
-        SecurityEvent(
-            event_type=EventType.ADMIN_PASSWORD_RESET,
-            severity=Severity.WARNING,
-            outcome="success",
-            actor_id=str(current_user.id),
-            actor_email=current_user.email,
-            actor_role=current_user.role.value,
-            target_id=str(user.id),
-            target_type="user",
-            detail=f"Password reset for {user.email}",
-        )
-    )
-    logger.warning("Admin %s reset password for user %s", current_user.email, user.email)
-    await audit(
-        current_user,
-        "admin.users.password_reset",
-        "user",
-        resource_id=str(user.id),
-        resource_name=user.email,
-    )
-
-    resp: dict[str, str] = {"message": f"Password reset for {user.email}"}
-    if req.generate:
-        resp["generated_password"] = new_password
-        resp["must_change_password"] = "true"
-    return resp
-
-
-async def _generate_unique_password(db: AsyncSession, length: int = 20, max_attempts: int = 10) -> str:
-    """Generate a secure password whose hash doesn't collide with any existing password hash."""
-    import os
-    import string
-
-    alphabet = string.ascii_letters + string.digits + string.punctuation
-    result = await db.execute(select(User.password_hash).where(User.password_hash.is_not(None)))
-    existing_hashes = {row[0] for row in result.all()}
-
-    for _ in range(max_attempts):
-        password = "".join(secrets.choice(alphabet) for _ in range(length))
-        # Check against all existing password hashes
-        salt = os.urandom(16)
-        key = hashlib.scrypt(password.encode(), salt=salt, n=16384, r=8, p=1, dklen=32)
-        candidate_hash = f"{salt.hex()}${key.hex()}"
-        if candidate_hash not in existing_hashes:
-            return password
-
-    # Astronomically unlikely to reach here, but be safe
-    return "".join(secrets.choice(alphabet) for _ in range(length))
-
-
-@router.delete("/users/{user_id}", status_code=204)
-async def delete_user(
-    user_id: uuid.UUID,
-    db: AsyncSession = Depends(get_db),
-    current_user: User = Depends(require_role(UserRole.admin)),
-):
-    """Admin deletes a user account and all associated data."""
-    if user_id == current_user.id:
-        raise HTTPException(status_code=400, detail="Cannot delete yourself")
-
-    stmt = select(User).where(User.id == user_id)
-    if current_user.org_id is not None:
-        stmt = stmt.where(User.org_id == current_user.org_id)
-    result = await db.execute(stmt)
-    user = result.scalar_one_or_none()
-    if not user:
-        raise HTTPException(status_code=404, detail="User not found")
-
-    # Prevent deleting the last admin/super_admin
-    if user.role in (UserRole.admin, UserRole.super_admin):
-        admin_count = await db.scalar(
-            select(func.count()).select_from(User).where(User.role.in_([UserRole.admin, UserRole.super_admin]))
-        )
-        if admin_count is not None and admin_count <= 1:
-            raise HTTPException(status_code=400, detail="Cannot delete the last admin")
-
-    logger.warning("Admin %s deleted user %s (%s)", current_user.email, user.email, user.id)
-    deleted_user_email = user.email
-    deleted_user_id = str(user.id)
-    await emit_security_event(
-        SecurityEvent(
-            event_type=EventType.USER_DELETED,
-            severity=Severity.WARNING,
-            outcome="success",
-            actor_id=str(current_user.id),
-            actor_email=current_user.email,
-            actor_role=current_user.role.value,
-            target_id=deleted_user_id,
-            target_type="user",
-            detail=f"Deleted user {deleted_user_email}",
-        )
-    )
-    await db.delete(user)
-    await db.commit()
-    await audit(
-        current_user,
-        "admin.users.delete",
-        "user",
-        resource_id=deleted_user_id,
-        resource_name=deleted_user_email,
-    )
-
-
-# ── Penalty & Weight Customization ──────────────────────
-
-
-@router.get("/penalties", response_model=list[dict])
-async def list_penalties(
-    db: AsyncSession = Depends(get_db),
-    current_user: User = Depends(require_role(UserRole.admin)),
-):
-    """List all penalty definitions."""
-    from models.scoring import PenaltyDefinition
-
-    result = await db.execute(
-        select(PenaltyDefinition).order_by(PenaltyDefinition.dimension, PenaltyDefinition.event_name)
-    )
-    penalties = [
-        {
-            "id": str(p.id),
-            "dimension": p.dimension.value,
-            "event_name": p.event_name,
-            "amount": p.amount,
-            "severity": p.severity.value,
-            "trigger_type": p.trigger_type.value,
-            "description": p.description,
-            "is_active": p.is_active,
-        }
-        for p in result.scalars().all()
-    ]
-    await audit(current_user, "admin.penalties.list", "penalty")
-    return penalties
-
-
-@router.put("/penalties/{penalty_id}", response_model=dict)
-async def update_penalty(
-    penalty_id: uuid.UUID,
-    req: dict,
-    db: AsyncSession = Depends(get_db),
-    current_user: User = Depends(require_role(UserRole.admin)),
-):
-    """Enable/disable or modify a penalty definition."""
-    from models.scoring import PenaltyDefinition
-
-    result = await db.execute(select(PenaltyDefinition).where(PenaltyDefinition.id == penalty_id))
-    penalty = result.scalar_one_or_none()
-    if not penalty:
-        raise HTTPException(status_code=404, detail="Penalty not found")
-
-    if "amount" in req:
-        penalty.amount = int(req["amount"])
-    if "is_active" in req:
-        penalty.is_active = bool(req["is_active"])
-    if "description" in req:
-        penalty.description = str(req["description"])
-
-    await db.commit()
-    await db.refresh(penalty)
-    await emit_security_event(
-        SecurityEvent(
-            event_type=EventType.PENALTY_WEIGHTS_MODIFIED,
-            severity=Severity.INFO,
-            outcome="success",
-            actor_id=str(current_user.id),
-            actor_email=current_user.email,
-            actor_role=current_user.role.value,
-            target_id=str(penalty_id),
-            target_type="penalty",
-            detail=f"Modified penalty {penalty.event_name}",
-        )
-    )
-    await audit(
-        current_user,
-        "admin.penalties.update",
-        "penalty",
-        resource_id=str(penalty_id),
-        resource_name=penalty.event_name,
-    )
-    return {
-        "id": str(penalty.id),
-        "event_name": penalty.event_name,
-        "amount": penalty.amount,
-        "is_active": penalty.is_active,
-    }
-
-
-@router.get("/weights", response_model=list[dict])
-async def list_weights(
-    db: AsyncSession = Depends(get_db),
-    current_user: User = Depends(require_role(UserRole.admin)),
-):
-    """List global dimension weights."""
-    from models.scoring import DEFAULT_DIMENSION_WEIGHTS, DimensionWeight
-
-    result = await db.execute(select(DimensionWeight).where(DimensionWeight.agent_id.is_(None)))
-    db_weights = {w.dimension.value: w.weight for w in result.scalars().all()}
-
-    # Merge with defaults
-    weights = []
-    for dim, default_weight in DEFAULT_DIMENSION_WEIGHTS.items():
-        weights.append(
-            {
-                "dimension": dim.value,
-                "weight": db_weights.get(dim.value, default_weight),
-                "is_custom": dim.value in db_weights,
-            }
-        )
-    await audit(current_user, "admin.weights.list", "weights")
-    return weights
-
-
-@router.put("/weights", response_model=dict)
-async def set_global_weights(
-    req: dict,
-    db: AsyncSession = Depends(get_db),
-    current_user: User = Depends(require_role(UserRole.admin)),
-):
-    """Set global dimension weights. Body: {dimension: weight, ...}"""
-    from models.scoring import DimensionWeight, ScoringDimension
-
-    updated = {}
-    for dim_name, weight in req.items():
-        try:
-            dim = ScoringDimension(dim_name)
-        except ValueError:
-            continue
-
-        result = await db.execute(
-            select(DimensionWeight).where(
-                DimensionWeight.agent_id.is_(None),
-                DimensionWeight.dimension == dim,
-            )
-        )
-        existing = result.scalar_one_or_none()
-        if existing:
-            existing.weight = float(weight)
-        else:
-            db.add(DimensionWeight(dimension=dim, weight=float(weight)))
-        updated[dim_name] = float(weight)
-
-    await db.commit()
-    await audit(
-        current_user,
-        "admin.weights.set_global",
-        "weights",
-        detail=json.dumps(updated),
-    )
-    return {"updated": updated}
-
-
-@router.put("/weights/agents/{agent_id}", response_model=dict)
-async def set_agent_weights(
-    agent_id: uuid.UUID,
-    req: dict,
-    db: AsyncSession = Depends(get_db),
-    current_user: User = Depends(require_role(UserRole.admin)),
-):
-    """Set per-agent dimension weights. Body: {dimension: weight, ...}"""
-    from models.scoring import DimensionWeight, ScoringDimension
-
-    updated = {}
-    for dim_name, weight in req.items():
-        try:
-            dim = ScoringDimension(dim_name)
-        except ValueError:
-            continue
-
-        result = await db.execute(
-            select(DimensionWeight).where(
-                DimensionWeight.agent_id == agent_id,
-                DimensionWeight.dimension == dim,
-            )
-        )
-        existing = result.scalar_one_or_none()
-        if existing:
-            existing.weight = float(weight)
-        else:
-            db.add(DimensionWeight(agent_id=agent_id, dimension=dim, weight=float(weight)))
-        updated[dim_name] = float(weight)
-
-    await db.commit()
-    await audit(
-        current_user,
-        "admin.weights.set_agent",
-        "weights",
-        resource_id=str(agent_id),
-        detail=json.dumps(updated),
-    )
-    return {"agent_id": str(agent_id), "updated": updated}
-
-
-# ── Canary Configuration ──────────────────────────────────
-
-# In-memory canary store (would be DB-backed in production)
-_canary_configs: dict[str, list[dict]] = {}  # agent_id -> list of canary configs
-_canary_reports: dict[str, list[dict]] = {}  # agent_id -> list of reports
-
-
-@router.post("/canaries", response_model=dict)
-async def create_canary(
-    req: dict,
-    current_user: User = Depends(require_role(UserRole.admin)),
-):
-    """Create a canary configuration for an agent."""
-    from services.eval.canary import CanaryConfig
-
-    agent_id = req.get("agent_id")
-    if not agent_id:
-        raise HTTPException(status_code=422, detail="agent_id required")
-
-    config = CanaryConfig(
-        agent_id=str(agent_id),
-        enabled=True,
-        canary_type=req.get("canary_type", "numeric"),
-        injection_point=req.get("injection_point", "tool_output"),
-        canary_value=req.get("canary_value", ""),
-        expected_behavior=req.get("expected_behavior", "flag_anomaly"),
-    )
-
-    if agent_id not in _canary_configs:
-        _canary_configs[agent_id] = []
-    _canary_configs[agent_id].append(config.model_dump())
-
-    await emit_security_event(
-        SecurityEvent(
-            event_type=EventType.CANARY_CREATED,
-            severity=Severity.INFO,
-            outcome="success",
-            actor_id=str(current_user.id),
-            actor_email=current_user.email,
-            actor_role=current_user.role.value,
-            target_id=str(agent_id),
-            target_type="canary",
-            detail=f"Canary type: {config.canary_type}",
-        )
-    )
-    await audit(
-        current_user,
-        "admin.canaries.create",
-        "canary",
-        resource_id=config.id,
-        detail=json.dumps({"agent_id": str(agent_id), "canary_type": config.canary_type}),
-    )
-    return {"id": config.id, "agent_id": agent_id, "canary_type": config.canary_type}
-
-
-@router.get("/canaries/{agent_id}", response_model=list[dict])
-async def list_canaries(
-    agent_id: str,
-    current_user: User = Depends(require_role(UserRole.admin)),
-):
-    """List canary configs for an agent."""
-    configs = _canary_configs.get(agent_id, [])
-    await audit(current_user, "admin.canaries.list", "canary", resource_id=agent_id)
-    return configs
-
-
-@router.get("/canaries/{agent_id}/reports", response_model=list[dict])
-async def list_canary_reports(
-    agent_id: str,
-    current_user: User = Depends(require_role(UserRole.admin)),
-):
-    """List canary reports with pass/fail stats."""
-    reports = _canary_reports.get(agent_id, [])
-    await audit(current_user, "admin.canaries.reports", "canary", resource_id=agent_id)
-    return reports
-
-
-@router.delete("/canaries/{canary_id}")
-async def delete_canary(
-    canary_id: str,
-    current_user: User = Depends(require_role(UserRole.admin)),
-):
-    """Remove a canary config."""
-    for _agent_id, configs in _canary_configs.items():
-        for i, config in enumerate(configs):
-            if config.get("id") == canary_id:
-                configs.pop(i)
-                await emit_security_event(
-                    SecurityEvent(
-                        event_type=EventType.CANARY_DELETED,
-                        severity=Severity.INFO,
-                        outcome="success",
-                        actor_id=str(current_user.id),
-                        actor_email=current_user.email,
-                        actor_role=current_user.role.value,
-                        target_id=canary_id,
-                        target_type="canary",
-                    )
-                )
-                await audit(current_user, "admin.canaries.delete", "canary", resource_id=canary_id)
-                return {"deleted": canary_id}
-    raise HTTPException(status_code=404, detail="Canary config not found")
-
-
-# ── Security Audit Log ──────────────────────────────────
-
-
-@router.get("/security-events")
-async def get_security_events(
-    event_type: str | None = None,
-    severity: str | None = None,
-    actor_email: str | None = None,
-    limit: int = 100,
-    offset: int = 0,
-    current_user: User = Depends(require_role(UserRole.admin)),
-):
-    """Query the security events audit log from ClickHouse."""
-    from services.clickhouse import _query
-
-    conditions = ["1 = 1"]
-    params: dict[str, str] = {}
-    if event_type:
-        conditions.append("event_type = {et:String}")
-        params["param_et"] = event_type
-    if severity:
-        conditions.append("severity = {sev:String}")
-        params["param_sev"] = severity
-    if actor_email:
-        conditions.append("actor_email = {ae:String}")
-        params["param_ae"] = actor_email
-
-    where = " AND ".join(conditions)
-    limit = min(max(int(limit), 1), 1000)
-    offset = max(int(offset), 0)
-    sql = (
-        f"SELECT * FROM security_events WHERE {where} ORDER BY timestamp DESC LIMIT {limit} OFFSET {offset} FORMAT JSON"
-    )
-    try:
-        r = await _query(sql, params)
-        r.raise_for_status()
-        data = r.json()
-        await audit(current_user, "admin.audit_log.view", "audit_log")
-        return {"events": data.get("data", []), "total": data.get("rows", 0)}
-    except Exception as e:
-        logger.warning("Audit log query failed: %s", e)
-        await audit(current_user, "admin.audit_log.view", "audit_log", detail="query_failed")
-        return {"events": [], "total": 0}
-
-
-@router.post("/resources/apply")
-async def apply_resources(
-    current_user: User = Depends(require_role(UserRole.admin)),
-    db: AsyncSession = Depends(get_db),
-):
-    """Re-apply resource tuning settings to ClickHouse without restart."""
-    from services.clickhouse import RESOURCE_SETTINGS_MAP, apply_resource_settings
-
-    result = await db.execute(select(EnterpriseConfig).where(EnterpriseConfig.key.like("resource.%")))
-    current = {cfg.key: cfg.value for cfg in result.scalars().all()}
-
-    await apply_resource_settings(overrides=current)
-
-    await emit_security_event(
-        SecurityEvent(
-            event_type=EventType.SETTING_CHANGED,
-            severity=Severity.WARNING,
-            outcome="success",
-            actor_id=str(current_user.id),
-            actor_email=current_user.email,
-            actor_role=current_user.role.value,
-            target_id="resource_settings",
-            target_type="setting",
-            detail=f"Applied resource settings: {list(current.keys())}",
-        )
-    )
-
-    applied_keys = [k for k in current if k in RESOURCE_SETTINGS_MAP]
-    await audit(
-        current_user,
-        "admin.resources.apply",
-        "resources",
-        detail=json.dumps(applied_keys),
-    )
-    return {
-        "applied": {k: current[k] for k in applied_keys},
-        "message": "ClickHouse resource settings applied",
-    }
-
-
-# ── Trace Privacy ──────────────────────────────────────────
-
-
-@router.get("/org/trace-privacy")
-async def get_trace_privacy(
-    db: AsyncSession = Depends(get_db),
-    current_user: User = Depends(require_role(UserRole.admin)),
-):
-    """Get the trace privacy setting for the current user's organization."""
-    if not current_user.org_id:
-        await audit(current_user, "admin.trace_privacy.view", "trace_privacy")
-        return {"trace_privacy": False}
-    result = await db.execute(select(Organization).where(Organization.id == current_user.org_id))
-    org = result.scalar_one_or_none()
-    if not org:
-        await audit(current_user, "admin.trace_privacy.view", "trace_privacy")
-        return {"trace_privacy": False}
-    await audit(current_user, "admin.trace_privacy.view", "trace_privacy", resource_id=str(org.id))
-    return {"trace_privacy": org.trace_privacy}
-
-
-@router.put("/org/trace-privacy")
-async def set_trace_privacy(
-    req: dict,
-    db: AsyncSession = Depends(get_db),
-    current_user: User = Depends(require_role(UserRole.admin)),
-):
-    """Toggle trace privacy for the current user's organization.
-
-    When enabled, all roles below super-admin can only see their own
-    traces.  Super-admins always retain full visibility.
-    """
-    enabled = bool(req.get("trace_privacy", False))
-
-    if not current_user.org_id:
-        raise HTTPException(status_code=400, detail="User has no organization")
-
-    result = await db.execute(select(Organization).where(Organization.id == current_user.org_id))
-    org = result.scalar_one_or_none()
-    if not org:
-        raise HTTPException(status_code=404, detail="Organization not found")
-
-    org.trace_privacy = enabled
-    await db.commit()
-    await db.refresh(org)
-    await emit_security_event(
-        SecurityEvent(
-            event_type=EventType.SETTING_CHANGED,
-            severity=Severity.WARNING,
-            outcome="success",
-            actor_id=str(current_user.id),
-            actor_email=current_user.email,
-            actor_role=current_user.role.value,
-            target_id=str(org.id),
-            target_type="organization",
-            detail=f"Trace privacy {'enabled' if enabled else 'disabled'}",
-        )
-    )
-    await audit(
-        current_user,
-        "admin.trace_privacy.update",
-        "trace_privacy",
-        resource_id=str(org.id),
-        detail=json.dumps({"enabled": enabled}),
-    )
-    return {"trace_privacy": org.trace_privacy}
-
-
-# ── Registered Agents Only ─────────────────────────────────
-
-
-@router.get("/org/registered-agents-only")
-async def get_registered_agents_only(
-    db: AsyncSession = Depends(get_db),
-    current_user: User = Depends(require_role(UserRole.user)),
-):
-    """Get the registered-agents-only setting for the current user's organization."""
-    if not current_user.org_id:
-        await audit(current_user, "admin.registered_agents_only.view", "registered_agents_only")
-        return {"registered_agents_only": False}
-    result = await db.execute(select(Organization).where(Organization.id == current_user.org_id))
-    org = result.scalar_one_or_none()
-    if not org:
-        await audit(current_user, "admin.registered_agents_only.view", "registered_agents_only")
-        return {"registered_agents_only": False}
-    await audit(current_user, "admin.registered_agents_only.view", "registered_agents_only", resource_id=str(org.id))
-    return {"registered_agents_only": org.registered_agents_only}
-
-
-@router.put("/org/registered-agents-only")
-async def set_registered_agents_only(
-    req: dict,
-    db: AsyncSession = Depends(get_db),
-    current_user: User = Depends(require_role(UserRole.super_admin)),
-):
-    """Toggle registered-agents-only mode for the current user's organization.
-
-    When enabled, only registered (active) agents are traced.
-    Unregistered agent telemetry is stored as metadata-only (no content).
-    """
-    enabled = bool(req.get("registered_agents_only", False))
-
-    if not current_user.org_id:
-        raise HTTPException(status_code=400, detail="User has no organization")
-
-    result = await db.execute(select(Organization).where(Organization.id == current_user.org_id))
-    org = result.scalar_one_or_none()
-    if not org:
-        raise HTTPException(status_code=404, detail="Organization not found")
-
-    org.registered_agents_only = enabled
-    await db.commit()
-    await db.refresh(org)
-    await emit_security_event(
-        SecurityEvent(
-            event_type=EventType.SETTING_CHANGED,
-            severity=Severity.WARNING,
-            outcome="success",
-            actor_id=str(current_user.id),
-            actor_email=current_user.email,
-            actor_role=current_user.role.value,
-            target_id=str(org.id),
-            target_type="organization",
-            detail=f"Registered-agents-only {'enabled' if enabled else 'disabled'}",
-        )
-    )
-    await audit(
-        current_user,
-        "admin.registered_agents_only.update",
-        "registered_agents_only",
-        resource_id=str(org.id),
-        detail=json.dumps({"enabled": enabled}),
-    )
-    # Invalidate registry cache so all server instances pick up the change immediately
-    from services.agent_registry_cache import invalidate as invalidate_registry_cache
-
-    await invalidate_registry_cache()
-    return {"registered_agents_only": org.registered_agents_only}
-
-
-@router.post("/cache/clear")
-async def clear_cache(current_user: User = Depends(require_role(UserRole.admin))):
-    """Clear all cached dashboard and OTEL responses."""
-    from services.cache import invalidate_all
-
-    deleted = await invalidate_all()
-    await audit(current_user, "admin.cache.clear", "cache", detail=json.dumps({"cleared": deleted}))
-    return {"cleared": deleted}
-
-
-@router.post("/fix-agent-org")
-async def fix_agent_org(
-    db: AsyncSession = Depends(get_db),
-    current_user: User = Depends(require_role(UserRole.admin)),
-):
-    """Fix agents missing owner_org_id by setting it from the creator's org."""
-    from models.agent import Agent, AgentVisibility
-
-    result = await db.execute(select(Agent).where(Agent.owner_org_id.is_(None)))
-    agents = result.scalars().all()
-    fixed = 0
-    for agent in agents:
-        creator = (await db.execute(select(User).where(User.id == agent.created_by))).scalar_one_or_none()
-        if creator and creator.org_id:
-            agent.owner_org_id = creator.org_id
-            agent.visibility = AgentVisibility.public
-            fixed += 1
-    await db.commit()
-    await audit(current_user, "admin.fix_agent_org", detail=json.dumps({"fixed": fixed}))
-    return {"fixed": fixed, "total_checked": len(agents)}
-
-
-# ── Data Retention ─────────────────────────────────────────
-
-
-@router.get("/org/retention")
-async def get_retention_config(
-    db: AsyncSession = Depends(get_db),
-    current_user: User = Depends(require_role(UserRole.admin)),
-) -> RetentionConfigResponse:
-    """Get the organization's data retention configuration."""
-
-    org = await _get_user_org(db, current_user)
-    global_retention = await ds.get_int("data.retention_days")
-    return RetentionConfigResponse(
-        retention_enabled=org.retention_enabled,
-        data_retention_days=org.data_retention_days,
-        score_retention_days=org.score_retention_days,
-        max_trace_count=org.max_trace_count,
-        global_retention_days=global_retention,
-    )
-
-
-@router.put("/org/retention")
-async def update_retention_config(
-    body: RetentionConfigUpdate,
-    db: AsyncSession = Depends(get_db),
-    current_user: User = Depends(require_role(UserRole.super_admin)),
-) -> RetentionConfigResponse:
-    """Update the organization's data retention configuration. Super admin only."""
-
-    global_retention = await ds.get_int("data.retention_days")
-    if body.data_retention_days is not None and global_retention > 0 and body.data_retention_days > global_retention:
-        raise HTTPException(
-            status_code=422,
-            detail=f"data_retention_days cannot exceed global ceiling of {global_retention} days",
-        )
-
-    org = await _get_user_org(db, current_user)
-    org.retention_enabled = body.retention_enabled
-    org.data_retention_days = body.data_retention_days
-    org.score_retention_days = body.score_retention_days
-    org.max_trace_count = body.max_trace_count
-    await db.commit()
-    await db.refresh(org)
-
-    await emit_security_event(
-        SecurityEvent(
-            event_type=EventType.SETTING_CHANGED,
-            severity=Severity.WARNING,
-            outcome="success",
-            actor_id=str(current_user.id),
-            actor_email=current_user.email,
-            actor_role=current_user.role.value,
-            target_id=str(org.id),
-            target_type="organization",
-            detail=f"Data retention {'enabled' if body.retention_enabled else 'disabled'}"
-            f" (days={body.data_retention_days}, scores={body.score_retention_days}, max={body.max_trace_count})",
-        )
-    )
-    await audit(
-        current_user,
-        "admin.retention.update",
-        "retention",
-        resource_id=str(org.id),
-        detail=json.dumps(
-            {
-                "retention_enabled": body.retention_enabled,
-                "data_retention_days": body.data_retention_days,
-                "score_retention_days": body.score_retention_days,
-                "max_trace_count": body.max_trace_count,
-            }
-        ),
-    )
-
-    return RetentionConfigResponse(
-        retention_enabled=org.retention_enabled,
-        data_retention_days=org.data_retention_days,
-        score_retention_days=org.score_retention_days,
-        max_trace_count=org.max_trace_count,
-        global_retention_days=global_retention,
-    )
-
-
-@router.get("/org/retention/preview")
-async def preview_retention(
-    days: int,
-    db: AsyncSession = Depends(get_db),
-    current_user: User = Depends(require_role(UserRole.super_admin)),
-):
-    """Preview approximate row counts that would be deleted for a given retention period."""
-    if days < 7:
-        raise HTTPException(status_code=422, detail="days must be >= 7")
-
-    from services.clickhouse import _query
-
-    org = await _get_user_org(db, current_user)
-    project_id = str(org.id)
-
-    counts = {}
-    tables = {
-        "traces": "SELECT count() as cnt FROM traces WHERE project_id = {pid:String} AND start_time < now() - INTERVAL {days:UInt32} DAY",
-        "spans": "SELECT count() as cnt FROM spans WHERE project_id = {pid:String} AND start_time < now() - INTERVAL {days:UInt32} DAY",
-        "scores": "SELECT count() as cnt FROM scores WHERE project_id = {pid:String} AND timestamp < now() - INTERVAL {days:UInt32} DAY",
-        "session_events": "SELECT count() as cnt FROM session_events WHERE project_id = {pid:String} AND timestamp < now() - INTERVAL {days:UInt32} DAY",
-    }
-
-    for table_name, sql in tables.items():
-        try:
-            resp = await _query(sql + " FORMAT JSON", {"param_pid": project_id, "param_days": str(days)})
-            if resp.status_code == 200:
-                data = resp.json().get("data", [{}])
-                counts[table_name] = int(data[0].get("cnt", 0)) if data else 0
-            else:
-                counts[table_name] = -1
-        except Exception:
-            counts[table_name] = -1
-
-    # Count insight reports from PostgreSQL
-    from datetime import UTC, datetime, timedelta
-
-    from models.agent import Agent
-    from models.insight_report import InsightReport, InsightReportStatus
-
-    score_cutoff = datetime.now(UTC) - timedelta(days=days * 2)
-    agent_ids = (await db.execute(select(Agent.id).where(Agent.owner_org_id == org.id))).scalars().all()
-    if agent_ids:
-        report_count = (
-            await db.execute(
-                select(func.count())
-                .select_from(InsightReport)
-                .where(
-                    InsightReport.agent_id.in_(agent_ids),
-                    InsightReport.completed_at < score_cutoff,
-                    InsightReport.status == InsightReportStatus.completed,
-                )
-            )
-        ).scalar() or 0
-    else:
-        report_count = 0
-
-    counts["insight_reports"] = report_count
-    counts["_note"] = "approximate; counts may be higher if a purge ran recently"
-    return counts
-
-
-@router.get("/org/retention/stats")
-async def get_retention_stats(
-    db: AsyncSession = Depends(get_db),
-    current_user: User = Depends(require_role(UserRole.admin)),
-):
-    """Get retention statistics for the dashboard widget."""
-    org = await _get_user_org(db, current_user)
-    if not org.retention_enabled:
-        return {
-            "retention_enabled": False,
-            "data_retention_days": org.data_retention_days,
-            "score_retention_days": org.score_retention_days,
-            "total_traces": 0,
-            "oldest_trace_age_days": 0,
-            "traces_expiring_7d": 0,
-            "next_purge_approx": None,
-        }
-
-    from services.clickhouse import _query
-
-    project_id = str(org.id)
-
-    # Get total traces and oldest trace
-    total_traces = 0
-    oldest_age_days = 0
-    try:
-        resp = await _query(
-            "SELECT count() as cnt, "
-            "if(cnt > 0, dateDiff('day', min(start_time), now()), 0) as age "
-            "FROM traces WHERE project_id = {pid:String} FORMAT JSON",
-            {"param_pid": project_id},
-        )
-        if resp.status_code == 200:
-            data = resp.json().get("data", [{}])
-            if data:
-                total_traces = int(data[0].get("cnt", 0))
-                oldest_age_days = int(data[0].get("age", 0)) if total_traces > 0 else 0
-    except Exception:
-        pass
-
-    # Get traces expiring in 7 days
-    traces_expiring = 0
-    if org.data_retention_days:
-        try:
-            cutoff_soon = org.data_retention_days - 7
-            if cutoff_soon > 0:
-                resp = await _query(
-                    "SELECT count() as cnt FROM traces "
-                    "WHERE project_id = {pid:String} "
-                    "AND start_time < now() - INTERVAL {days:UInt32} DAY "
-                    "FORMAT JSON",
-                    {"param_pid": project_id, "param_days": str(cutoff_soon)},
-                )
-                if resp.status_code == 200:
-                    data = resp.json().get("data", [{}])
-                    if data:
-                        traces_expiring = int(data[0].get("cnt", 0))
-        except Exception:
-            pass
-
-    return {
-        "retention_enabled": True,
-        "data_retention_days": org.data_retention_days,
-        "score_retention_days": org.score_retention_days,
-        "total_traces": total_traces,
-        "oldest_trace_age_days": oldest_age_days,
-        "traces_expiring_7d": traces_expiring,
-        "next_purge_approx": "Every 6 hours (01:30, 07:30, 13:30, 19:30 UTC)",
-    }
-
-
-@router.get("/org/retention/warnings")
-async def get_retention_warnings(
-    db: AsyncSession = Depends(get_db),
-    current_user: User = Depends(require_role(UserRole.admin)),
-):
-    """Get agents with unanalyzed traces approaching expiry."""
-    org = await _get_user_org(db, current_user)
-    if not org.retention_enabled or not org.data_retention_days:
-        return {"warnings": [], "retention_days": org.data_retention_days, "retention_enabled": org.retention_enabled}
-
-    from models.agent import Agent
-    from models.insight_report import InsightReport, InsightReportStatus
-
-    # Get all agents for this org
-    agents_result = await db.execute(select(Agent.id, Agent.name).where(Agent.owner_org_id == org.id))
-    agents = agents_result.all()
-    if not agents:
-        return {"warnings": [], "retention_days": org.data_retention_days, "retention_enabled": True}
-
-    # Get latest completed report per agent
-    latest_reports = {}
-    for agent_id, _ in agents:
-        report = (
-            await db.execute(
-                select(InsightReport.completed_at)
-                .where(
-                    InsightReport.agent_id == agent_id,
-                    InsightReport.status == InsightReportStatus.completed,
-                )
-                .order_by(InsightReport.completed_at.desc())
-                .limit(1)
-            )
-        ).scalar_one_or_none()
-        latest_reports[agent_id] = report
-
-    # Find agents without a recent report covering the retention window
-    from datetime import UTC, datetime, timedelta
-
-    retention_start = datetime.now(UTC) - timedelta(days=org.data_retention_days)
-    warnings = []
-    for agent_id, agent_name in agents:
-        last_report = latest_reports.get(agent_id)
-        if last_report is None or last_report < retention_start:
-            warnings.append(
-                {
-                    "agent_id": str(agent_id),
-                    "agent_name": agent_name or "Unnamed Agent",
-                    "traces_expiring_soon": 0,
-                    "last_insight_report": last_report.isoformat() if last_report else None,
-                }
-            )
-
-    return {"warnings": warnings, "retention_days": org.data_retention_days, "retention_enabled": True}
-
-
-async def _get_user_org(db: AsyncSession, user: User) -> Organization:
-    """Get the user's org. Raises 400 if user has no org."""
-    if not user.org_id:
-        raise HTTPException(status_code=400, detail="User has no organization")
-    org = (await db.execute(select(Organization).where(Organization.id == user.org_id))).scalar_one_or_none()
-    if not org:
-        raise HTTPException(status_code=404, detail="Organization not found")
-    return org
